+++ R/CalcMaxMismatch.R
@@ -27,8 +27,8 @@
 #'   'version1.1'. Ignored if \code{Err} is a matrix. See \code{\link{ErrToM}}.
 #' @param qntl quantile of binomial distribution to be used as the maximum, of
 #'   individual-level probability. For a desired dataset-level probability
-#'   quantile \eqn{Q}, use \eqn{qntl = Q^(1/N)}, where \eqn{N} is the number of
-#'   individuals.
+#'   quantile \eqn{Q}, use \code{qntl}\eqn{= Q^{(1/N)}}, where \eqn{N} is
+#'   the number of individuals.
 #'
 #' @return A vector with three integers:
 #'  \item{DUP}{Maximum number of differences between 2 samples from the